Major revisions thanks to professor and friends' valuable feedback
Expand Cache on NUMA Systems section to introduce cache coherency give a
better idea where cache is situated (which I am still a little fuzzy
about).

Migrated the explanation of synchronous access to the locks and lock
contention section because it fits well there.

Much work still needs to be done to expand and clarify the methods
section.